Updated Memory Management section
